refactor(main): reorg normalizations and validations
moves the ruleset validations closer to main (and separate from the validation code

moves the options validation/ normalization into a separate folder
