+++ src/main/index.js
@@ -2,14 +2,15 @@
 
 const extract          = require("../extract");
 const meta             = require("../extract/transpile/meta");
-const readRuleSet      = require("../validate/readRuleSet");
+const validateRuleSet  = require("./ruleSet/validate");
+const normalizeRuleSet = require("./ruleSet/normalize");
 const reportHtml       = require("../report/htmlReporter");
 const reportJson       = require("../report/jsonReporter");
 const reportDot        = require("../report/dotReporter");
 const reportCsv        = require("../report/csvReporter");
 const reportErr        = require("../report/errReporter");
-const validateOptions  = require("./validateOptions");
-const normalizeOptions = require("./normalizeOptions");
+const validateOptions  = require("./options/validate");
+const normalizeOptions = require("./options/normalize");
 
 const TYPE2REPORTER      = {
     "json" : reportJson,
@@ -30,7 +31,7 @@ const TYPE2REPORTER      = {
  *                Default false.
  *  ruleSet     : An object (or JSON string) containing the rules to validate
  *                against. The rules should adhere to the
- *                [ruleset schema](../src/extract/validate/jsonschema.json)
+ *                [ruleset schema](../src/main/ruleSet/jsonschema.json)
  *                The function with throw an Error when either
  *                - the passed ruleSet violates that schema or
  *                - it contains an 'unsafe' (= potentially super slow running)
@@ -74,7 +75,11 @@ exports.cruise = (pFileDirArray, pOptions) => {
     );
 
     if (Boolean(pOptions.ruleSet)){
-        pOptions.ruleSet = readRuleSet(pOptions.ruleSet);
+        pOptions.ruleSet = normalizeRuleSet(
+            validateRuleSet(
+                typeof pOptions.ruleSet === 'object' ? pOptions.ruleSet : JSON.parse(pOptions.ruleSet)
+            )
+        );
     }
 
     return extract(
