+++ src/main/options/normalize.js
