+++ src/main/options/validate.js
