+++ test/main/options/validate.spec.js
@@ -1,6 +1,6 @@
 "use strict";
 const expect          = require('chai').expect;
-const validateOptions = require('../../src/main/validateOptions');
+const validateOptions = require('../../../src/main/options/validate');
 
 describe("validateOptions", () => {
