+++ test/main/ruleSet/validate.spec.js
@@ -1,13 +1,15 @@
 "use strict";
-const expect      = require('chai').expect;
-const readRuleSet = require('../../src/validate/readRuleSet');
-const fs          = require('fs');
+const expect   = require('chai').expect;
+const validate = require('../../../src/main/ruleSet/validate');
+const fs       = require('fs');
 
 describe("ruleSetReader", () => {
     it("bails out on scary regexps in paths", () => {
         try {
-            readRuleSet(
-                fs.readFileSync("./test/validate/fixtures/rules.scary-regex.json", 'utf8')
+            validate(
+                JSON.parse(
+                    fs.readFileSync("./test/validate/fixtures/rules.scary-regex.json", 'utf8')
+                )
             );
             expect("not to be here").to.equal("still here, though");
         } catch (e) {
@@ -19,8 +21,10 @@ describe("ruleSetReader", () => {
 
     it("bails out on scary regexps in pathNots", () => {
         try {
-            readRuleSet(
-                fs.readFileSync("./test/validate/fixtures/rules.scary-regex-in-pathnot.json", 'utf8')
+            validate(
+                JSON.parse(
+                    fs.readFileSync("./test/validate/fixtures/rules.scary-regex-in-pathnot.json", 'utf8')
+                )
             );
             expect("not to be here").to.equal("still here, though");
         } catch (e) {
@@ -32,7 +36,7 @@ describe("ruleSetReader", () => {
 
     it("barfs on an invalid rules file", () => {
         try {
-            readRuleSet(
+            validate(
                 fs.readFileSync("./test/validate/fixtures/rules.not-a-valid-rulesfile.json")
             );
             expect("not to be here").to.equal("still here, though");
@@ -44,6 +48,6 @@ describe("ruleSetReader", () => {
     });
 
     it("also accept javascript objects", () => {
-        expect(readRuleSet({})).to.deep.equal({});
+        expect(validate({})).to.deep.equal({});
     });
 });
